Cast int to long int: 32 long:64

+++ src/leetcode/SqrtOfX.java
@@ -7,25 +7,25 @@ package leetcode;
  */
 public class SqrtOfX {
     //This is NOT an easy one. Int overflow needs to be considered!!!
-    //Cast int to double makes it easier.
+    //Cast int to long makes it easier.
     public int sqrt(int x) {
         // Start typing your Java solution below
         // DO NOT write main() function
-        return sqrt((double) x, 0, x);
+        return sqrt((long) x, 0, x);
     }
 
-    private int sqrt(double x, int start, int end) {
+    private int sqrt(long x, int start, int end) {
         int mid = start + (end - start) / 2;
-        if ((double) mid * mid == x) {
+        if ((long) mid * mid == x) {
             return mid;
-        } else if ((double) mid * mid < x) {
-            if (Math.abs((double) mid * mid - x) < Math.abs((double) (mid + 1) * (mid + 1) - x)) {
+        } else if ((long) mid * mid < x) {
+            if (Math.abs((long) mid * mid - x) < Math.abs((long) (mid + 1) * (mid + 1) - x)) {
                 return mid;
             } else {
                 return sqrt(x, mid + 1, end);
             }
         } else {
-            if (Math.abs((double) mid * mid - x) < Math.abs((double) (mid - 1) * (mid - 1) - x)) {
+            if (Math.abs((long) mid * mid - x) < Math.abs((long) (mid - 1) * (mid - 1) - x)) {
                 return mid - 1;
             } else {
                 return sqrt(x, start, mid - 1);
